backend/requirements.txt
click==6.7
Flask==0.12.2
Flask-SQLAlchemy==2.1
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
PyMySQL==0.8.0
SQLAlchemy==1.2.5
Werkzeug==0.14.1
PyJWT==1.6.0
